refactor(parser): split parser/mod.rs and ast.rs into submodules. Closes #199
- Split parser/mod.rs (1607 lines) into: mod.rs (324), agent_parser.rs (473),
  step_parser.rs (323), workflow_parser.rs (344), type_parser.rs (119),
  import_parser.rs (75)
- Split ast.rs (772 lines) into: mod.rs (49), agent.rs (79), workflow.rs (221),
  provider.rs (27), types.rs (31), value.rs (72), import.rs (23)
- All files under 400 lines
- Also addresses #202 by refactoring parse_step (removed #[allow(clippy::too_many_lines)])
  and parse_workflow into smaller helper functions
