--- src/ast.rs
@@ -1,772 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-/// Byte-offset span in the source file.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Span {
-    pub start: usize,
-    pub end: usize,
-}
-
-impl Span {
-    pub fn new(start: usize, end: usize) -> Self {
-        Self { start, end }
-    }
-}
-
-/// A value expression used in configuration fields.
-///
-/// Supports literal strings and function calls like `env("VAR_NAME")`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum ValueExpr {
-    /// A plain string or identifier value.
-    Literal(String),
-    /// An environment variable reference: `env("VAR_NAME")`.
-    EnvRef { var_name: String, span: Span },
-}
-
-/// Error from resolving a `ValueExpr`.
-#[derive(Debug, Clone, PartialEq)]
-pub enum ResolveError {
-    /// An environment variable was not found.
-    EnvVarNotSet(String),
-}
-
-impl std::fmt::Display for ResolveError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::EnvVarNotSet(var) => write!(f, "environment variable '{var}' is not set"),
-        }
-    }
-}
-
-impl std::error::Error for ResolveError {}
-
-impl ValueExpr {
-    /// Resolve to a plain string value using the provided env lookup function.
-    ///
-    /// For `Literal`, returns the string directly. For `EnvRef`, calls
-    /// `env_lookup` with the variable name.
-    pub fn resolve_with<F>(&self, env_lookup: F) -> Result<String, ResolveError>
-    where
-        F: Fn(&str) -> Option<String>,
-    {
-        match self {
-            Self::Literal(s) => Ok(s.clone()),
-            Self::EnvRef { var_name, .. } => {
-                env_lookup(var_name).ok_or_else(|| ResolveError::EnvVarNotSet(var_name.clone()))
-            }
-        }
-    }
-
-    /// Resolve using `std::env::var`. Convenience wrapper around `resolve_with`.
-    pub fn resolve(&self) -> Result<String, ResolveError> {
-        self.resolve_with(|name| std::env::var(name).ok())
-    }
-
-    /// Return the literal string value if this is a `Literal`.
-    pub fn as_literal(&self) -> Option<&str> {
-        match self {
-            Self::Literal(s) => Some(s),
-            Self::EnvRef { .. } => None,
-        }
-    }
-
-    /// Return a display-friendly string for this value.
-    /// For `Literal`, returns the string. For `EnvRef`, returns the var name.
-    pub fn display_value(&self) -> &str {
-        match self {
-            Self::Literal(s) => s,
-            Self::EnvRef { var_name, .. } => var_name,
-        }
-    }
-}
-
-/// A monetary cap constraint on a capability (`up to $<amount>`).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "type")]
-pub enum Constraint {
-    MonetaryCap { amount: u64, currency: String },
-}
-
-/// A single tool capability, e.g. `zendesk.read_ticket` or `zendesk.refund up to $50`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Capability {
-    pub namespace: String,
-    pub action: String,
-    pub constraint: Option<Constraint>,
-    pub span: Span,
-}
-
-/// A spending budget, e.g. `budget: $0.03 per ticket`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Budget {
-    pub amount: u64,
-    pub currency: String,
-    pub unit: String,
-    pub span: Span,
-}
-
-/// A single `agent <name> { ... }` definition.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct AgentDef {
-    pub name: String,
-    pub model: Option<ValueExpr>,
-    pub can: Vec<Capability>,
-    pub cannot: Vec<Capability>,
-    pub budget: Option<Budget>,
-    pub guardrails: Option<GuardrailsDef>,
-    pub span: Span,
-}
-
-// ---------------------------------------------------------------------------
-// Defaults types
-// ---------------------------------------------------------------------------
-
-/// A `defaults { model: ..., budget: ... }` block providing project-level defaults.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct DefaultsDef {
-    pub model: Option<ValueExpr>,
-    pub budget: Option<Budget>,
-    pub span: Span,
-}
-
-// ---------------------------------------------------------------------------
-// Guardrails types
-// ---------------------------------------------------------------------------
-
-/// A `guardrails { ... }` block containing named sections.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct GuardrailsDef {
-    pub sections: Vec<GuardrailSection>,
-    pub span: Span,
-}
-
-/// A named section within guardrails, e.g. `output_filter { pii_detection: redact }`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct GuardrailSection {
-    pub name: String,
-    pub rules: Vec<GuardrailRule>,
-    pub span: Span,
-}
-
-/// A key-value rule within a guardrail section, e.g. `pii_detection: redact`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct GuardrailRule {
-    pub key: String,
-    pub value: String,
-    pub span: Span,
-}
-
-// ---------------------------------------------------------------------------
-// Provider types
-// ---------------------------------------------------------------------------
-
-/// A `provider <name> { ... }` definition.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ProviderDef {
-    pub name: String,
-    /// The default model for this provider (e.g. `claude-haiku`, `gpt-4o`).
-    pub model: Option<ValueExpr>,
-    /// API key or credential reference (e.g. `env("ANTHROPIC_KEY")`).
-    pub key: Option<ValueExpr>,
-    pub span: Span,
-}
-
-// ---------------------------------------------------------------------------
-// Tool types
-// ---------------------------------------------------------------------------
-
-/// A tool block: `tool zendesk { endpoint: "https://..." }`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ToolDef {
-    pub name: String,
-    /// The endpoint URL for this tool.
-    pub endpoint: Option<ValueExpr>,
-    /// Optional provider type (e.g. `rest_api`, `mcp`).
-    pub provider: Option<ValueExpr>,
-    /// Optional API key or credential reference.
-    pub key: Option<ValueExpr>,
-    pub span: Span,
-}
-
-// ---------------------------------------------------------------------------
-// Workflow types
-// ---------------------------------------------------------------------------
-
-/// A type expression constraining a field's value.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum TypeExpr {
-    /// A built-in or user-defined named type (e.g. `string`, `int`, `MyType`).
-    Named { name: String, array: bool },
-    /// A union of allowed values: `one of [billing, technical, general]`.
-    OneOf {
-        variants: Vec<String>,
-        span: Span,
-    },
-    /// A numeric range: `1..10` or `0.0..1.0`.
-    Range {
-        min: String,
-        max: String,
-    },
-}
-
-/// A field within a `type` definition.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct TypeField {
-    pub name: String,
-    pub type_expr: TypeExpr,
-    pub span: Span,
-}
-
-/// A `type Name { ... }` definition.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct TypeDef {
-    pub name: String,
-    pub fields: Vec<TypeField>,
-    pub span: Span,
-}
-
-/// A single condition in an `auto resolve when` block.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum AutoResolveCondition {
-    /// A comparison: `confidence > 90%`.
-    Comparison(WhenComparison),
-    /// A membership check: `action is one of [order_status, tracking]`.
-    IsOneOf {
-        field: String,
-        variants: Vec<String>,
-    },
-}
-
-/// An `auto resolve when { ... }` block for autonomous agent actions.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct AutoResolveBlock {
-    pub conditions: Vec<AutoResolveCondition>,
-    pub span: Span,
-}
-
-/// Backoff strategy for retry policies.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum BackoffStrategy {
-    Exponential,
-    Linear,
-    Fixed,
-}
-
-/// Action to take after retries are exhausted.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum FailureAction {
-    /// Escalate to a human or higher-level agent.
-    Escalate,
-    /// Execute a named step.
-    Step(String),
-}
-
-/// A retry policy: `on failure: retry N strategy then action`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct RetryPolicy {
-    pub max_retries: u32,
-    pub backoff: BackoffStrategy,
-    pub then: FailureAction,
-}
-
-/// A comparison operator in a `when` expression.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum CompareOp {
-    Lt,
-    Gt,
-    LtEq,
-    GtEq,
-}
-
-/// A value in a `when` comparison (right-hand side).
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum WhenValue {
-    /// A percentage like `70%`.
-    Percent(String),
-    /// A currency amount like `$50`.
-    Currency { symbol: char, amount: u64 },
-    /// A plain number.
-    Number(String),
-    /// An identifier reference.
-    Ident(String),
-}
-
-/// A single comparison in a `when` expression: `field op value`.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct WhenComparison {
-    pub field: String,
-    pub op: CompareOp,
-    pub value: WhenValue,
-}
-
-/// A `when` expression combining comparisons with boolean logic.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum WhenExpr {
-    Comparison(WhenComparison),
-    And(Vec<WhenExpr>),
-    Or(Vec<WhenExpr>),
-}
-
-/// A `parallel { step a {...} step b {...} }` block for concurrent execution.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ParallelBlock {
-    pub steps: Vec<StepDef>,
-    pub span: Span,
-}
-
-/// A single arm in a `route on` block.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct RouteArm {
-    /// The match pattern: a literal value or `_` for wildcard/default.
-    pub pattern: RoutePattern,
-    /// The step to execute if this arm matches.
-    pub step: StepDef,
-    pub span: Span,
-}
-
-/// A match pattern in a route arm.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum RoutePattern {
-    /// Match a specific value.
-    Value(String),
-    /// Wildcard: match anything not matched above.
-    Wildcard,
-}
-
-/// A `route on <expr> { ... }` block for pattern-matched routing.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct RouteBlock {
-    /// Dot-separated field path to match on (e.g. `classify.category`).
-    pub field_path: String,
-    /// Match arms in order.
-    pub arms: Vec<RouteArm>,
-    pub span: Span,
-}
-
-/// An import declaration.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum ImportDef {
-    /// Named imports: `import { name1, name2 } from "./path.rein"`.
-    Named {
-        names: Vec<String>,
-        source: String,
-        span: Span,
-    },
-    /// Glob import: `import all from "./dir/"`.
-    Glob {
-        source: String,
-        span: Span,
-    },
-    /// Registry import: `import from @scope/name`.
-    Registry {
-        scope: String,
-        name: String,
-        span: Span,
-    },
-}
-
-/// How a condition is evaluated against agent output.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "mode", content = "value", rename_all = "snake_case")]
-pub enum ConditionMatcher {
-    /// Exact value match (case-insensitive, word-boundary aware).
-    Equals(String),
-    /// Substring containment (case-insensitive).
-    Contains(String),
-    /// Regular expression match.
-    Regex(String),
-    /// JSON path extraction and comparison (`path=expected`).
-    JsonPath {
-        path: String,
-        expected: String,
-    },
-}
-
-/// How a workflow stage routes to the next stage.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum RouteRule {
-    /// Always go to the next stage in sequence.
-    Next,
-    /// Route based on a condition in the agent's output.
-    Conditional {
-        field: String,
-        matcher: ConditionMatcher,
-        then_stage: String,
-        else_stage: Option<String>,
-    },
-}
-
-/// A single stage in a workflow pipeline.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Stage {
-    /// Name of this stage (used for routing references).
-    pub name: String,
-    /// Name of the agent to run at this stage.
-    pub agent: String,
-    /// How to route after this stage completes.
-    pub route: RouteRule,
-    pub span: Span,
-}
-
-/// Execution mode for a group of stages.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ExecutionMode {
-    Sequential,
-    Parallel,
-}
-
-/// A `step <name> { agent: <name> goal: <text> }` definition within a workflow.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct StepDef {
-    pub name: String,
-    /// The agent to execute for this step.
-    pub agent: String,
-    /// Natural language goal describing what the step should accomplish.
-    pub goal: Option<String>,
-    /// Output type constraints (e.g. `category: one of [billing, technical]`).
-    pub output_constraints: Vec<(String, TypeExpr)>,
-    /// Optional guard condition: `when: confidence < 70%`.
-    pub when: Option<WhenExpr>,
-    /// Optional retry policy: `on failure: retry 3 exponential then escalate`.
-    pub on_failure: Option<RetryPolicy>,
-    pub span: Span,
-}
-
-/// A `workflow <name> { ... }` definition.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct WorkflowDef {
-    pub name: String,
-    /// What triggers this workflow (e.g. `incoming_ticket`).
-    pub trigger: String,
-    /// Ordered list of stages.
-    pub stages: Vec<Stage>,
-    /// Named step blocks (`step <name> { ... }`).
-    pub steps: Vec<StepDef>,
-    /// Route-on blocks for pattern-matched routing.
-    pub route_blocks: Vec<RouteBlock>,
-    /// Parallel execution blocks.
-    pub parallel_blocks: Vec<ParallelBlock>,
-    /// Auto resolve conditions.
-    pub auto_resolve: Option<AutoResolveBlock>,
-    /// Default execution mode.
-    pub mode: ExecutionMode,
-    pub span: Span,
-}
-
-impl WorkflowDef {
-    /// Find a stage by name within this workflow.
-    pub fn find_stage(&self, name: &str) -> Option<&Stage> {
-        self.stages.iter().find(|s| s.name == name)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Top-level file
-// ---------------------------------------------------------------------------
-
-/// Top-level parsed file — provider, agent, and workflow definitions.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ReinFile {
-    pub imports: Vec<ImportDef>,
-    pub defaults: Option<DefaultsDef>,
-    pub providers: Vec<ProviderDef>,
-    pub tools: Vec<ToolDef>,
-    pub agents: Vec<AgentDef>,
-    pub workflows: Vec<WorkflowDef>,
-    pub types: Vec<TypeDef>,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn dummy_span() -> Span {
-        Span::new(0, 1)
-    }
-
-    #[test]
-    fn span_serializes_to_json() {
-        let span = Span::new(10, 20);
-        let json = serde_json::to_string(&span).unwrap();
-        assert_eq!(json, r#"{"start":10,"end":20}"#);
-    }
-
-    #[test]
-    fn constraint_monetary_cap_serializes() {
-        let c = Constraint::MonetaryCap {
-            amount: 5000,
-            currency: "USD".to_string(),
-        };
-        let json = serde_json::to_value(&c).unwrap();
-        assert_eq!(json["type"], "MonetaryCap");
-        assert_eq!(json["amount"], 5000);
-        assert_eq!(json["currency"], "USD");
-    }
-
-    #[test]
-    fn capability_with_constraint_serializes() {
-        let cap = Capability {
-            namespace: "zendesk".to_string(),
-            action: "refund".to_string(),
-            constraint: Some(Constraint::MonetaryCap {
-                amount: 5000,
-                currency: "USD".to_string(),
-            }),
-            span: dummy_span(),
-        };
-        let json = serde_json::to_value(&cap).unwrap();
-        assert_eq!(json["namespace"], "zendesk");
-        assert_eq!(json["action"], "refund");
-        assert_eq!(json["constraint"]["amount"], 5000);
-    }
-
-    #[test]
-    fn capability_without_constraint_serializes() {
-        let cap = Capability {
-            namespace: "zendesk".to_string(),
-            action: "read_ticket".to_string(),
-            constraint: None,
-            span: dummy_span(),
-        };
-        let json = serde_json::to_value(&cap).unwrap();
-        assert_eq!(json["constraint"], serde_json::Value::Null);
-    }
-
-    #[test]
-    fn budget_serializes() {
-        let b = Budget {
-            amount: 3,
-            currency: "USD".to_string(),
-            unit: "ticket".to_string(),
-            span: dummy_span(),
-        };
-        let json = serde_json::to_value(&b).unwrap();
-        assert_eq!(json["amount"], 3);
-        assert_eq!(json["currency"], "USD");
-        assert_eq!(json["unit"], "ticket");
-    }
-
-    #[test]
-    fn agent_def_full_serializes() {
-        let agent = AgentDef {
-            name: "support_triage".to_string(),
-            model: Some(ValueExpr::Literal("anthropic".into())),
-            can: vec![Capability {
-                namespace: "zendesk".to_string(),
-                action: "read_ticket".to_string(),
-                constraint: None,
-                span: dummy_span(),
-            }],
-            cannot: vec![Capability {
-                namespace: "zendesk".to_string(),
-                action: "delete_ticket".to_string(),
-                constraint: None,
-                span: dummy_span(),
-            }],
-            budget: Some(Budget {
-                amount: 3,
-                currency: "USD".to_string(),
-                unit: "ticket".to_string(),
-                span: dummy_span(),
-            }),
-            guardrails: None,
-            span: dummy_span(),
-        };
-        let json = serde_json::to_value(&agent).unwrap();
-        assert_eq!(json["name"], "support_triage");
-        assert_eq!(json["model"], "anthropic");
-        assert_eq!(json["can"].as_array().unwrap().len(), 1);
-        assert_eq!(json["cannot"].as_array().unwrap().len(), 1);
-        assert_eq!(json["budget"]["unit"], "ticket");
-    }
-
-    #[test]
-    fn rein_file_roundtrips_via_json() {
-        let file = ReinFile {
-            imports: vec![],
-            defaults: None,
-            providers: vec![],
-            tools: vec![],
-            agents: vec![AgentDef {
-                name: "bot".to_string(),
-                model: None,
-                can: vec![],
-                cannot: vec![],
-                budget: None,
-                guardrails: None,
-                span: dummy_span(),
-            }],
-            workflows: vec![],
-            types: vec![],
-        };
-        let json = serde_json::to_string(&file).unwrap();
-        let decoded: ReinFile = serde_json::from_str(&json).unwrap();
-        assert_eq!(file, decoded);
-    }
-
-    #[test]
-    fn agent_def_minimal_model_none() {
-        let agent = AgentDef {
-            name: "minimal".to_string(),
-            model: None,
-            can: vec![],
-            cannot: vec![],
-            budget: None,
-            guardrails: None,
-            span: dummy_span(),
-        };
-        let json = serde_json::to_value(&agent).unwrap();
-        assert_eq!(json["model"], serde_json::Value::Null);
-        assert!(json["can"].as_array().unwrap().is_empty());
-    }
-
-    // ── Workflow types ─────────────────────────────────────────────────
-
-    #[test]
-    fn stage_serializes() {
-        let stage = Stage {
-            name: "triage".to_string(),
-            agent: "support_triage".to_string(),
-            route: RouteRule::Next,
-            span: dummy_span(),
-        };
-        let json = serde_json::to_value(&stage).unwrap();
-        assert_eq!(json["name"], "triage");
-        assert_eq!(json["agent"], "support_triage");
-        assert_eq!(json["route"]["type"], "next");
-    }
-
-    #[test]
-    fn conditional_route_serializes() {
-        let route = RouteRule::Conditional {
-            field: "sentiment".to_string(),
-            matcher: ConditionMatcher::Equals("negative".to_string()),
-            then_stage: "escalate".to_string(),
-            else_stage: Some("respond".to_string()),
-        };
-        let json = serde_json::to_value(&route).unwrap();
-        assert_eq!(json["type"], "conditional");
-        assert_eq!(json["field"], "sentiment");
-        assert_eq!(json["then_stage"], "escalate");
-    }
-
-    #[test]
-    fn workflow_def_serializes() {
-        let workflow = WorkflowDef {
-            name: "support_pipeline".to_string(),
-            trigger: "incoming_ticket".to_string(),
-            stages: vec![
-                Stage {
-                    name: "triage".to_string(),
-                    agent: "support_triage".to_string(),
-                    route: RouteRule::Next,
-                    span: dummy_span(),
-                },
-                Stage {
-                    name: "respond".to_string(),
-                    agent: "responder".to_string(),
-                    route: RouteRule::Next,
-                    span: dummy_span(),
-                },
-            ],
-            steps: vec![],
-            route_blocks: vec![],
-            parallel_blocks: vec![],
-            auto_resolve: None,
-            mode: ExecutionMode::Sequential,
-            span: dummy_span(),
-        };
-        let json = serde_json::to_value(&workflow).unwrap();
-        assert_eq!(json["name"], "support_pipeline");
-        assert_eq!(json["trigger"], "incoming_ticket");
-        assert_eq!(json["stages"].as_array().unwrap().len(), 2);
-        assert_eq!(json["mode"], "sequential");
-    }
-
-    #[test]
-    fn workflow_roundtrips_via_json() {
-        let workflow = WorkflowDef {
-            name: "test".to_string(),
-            trigger: "event".to_string(),
-            stages: vec![],
-            steps: vec![],
-            route_blocks: vec![],
-            parallel_blocks: vec![],
-            auto_resolve: None,
-            mode: ExecutionMode::Parallel,
-            span: dummy_span(),
-        };
-        let json = serde_json::to_string(&workflow).unwrap();
-        let decoded: WorkflowDef = serde_json::from_str(&json).unwrap();
-        assert_eq!(workflow, decoded);
-    }
-
-    #[test]
-    fn rein_file_with_workflows_roundtrips() {
-        let file = ReinFile {
-            imports: vec![],
-            defaults: None,
-            providers: vec![],
-            tools: vec![],
-            agents: vec![],
-            workflows: vec![WorkflowDef {
-                name: "pipeline".to_string(),
-                trigger: "event".to_string(),
-                stages: vec![],
-                steps: vec![],
-            route_blocks: vec![],
-            parallel_blocks: vec![],
-            auto_resolve: None,
-                mode: ExecutionMode::Sequential,
-                span: dummy_span(),
-            }],
-            types: vec![],
-        };
-        let json = serde_json::to_string(&file).unwrap();
-        let decoded: ReinFile = serde_json::from_str(&json).unwrap();
-        assert_eq!(file, decoded);
-    }
-
-    #[test]
-    fn workflow_def_find_stage() {
-        let wf = WorkflowDef {
-            name: "test".to_string(),
-            trigger: "event".to_string(),
-            stages: vec![
-                Stage {
-                    name: "a".to_string(),
-                    agent: "agent_a".to_string(),
-                    route: RouteRule::Next,
-                    span: dummy_span(),
-                },
-                Stage {
-                    name: "b".to_string(),
-                    agent: "agent_b".to_string(),
-                    route: RouteRule::Next,
-                    span: dummy_span(),
-                },
-            ],
-            steps: vec![],
-            route_blocks: vec![],
-            parallel_blocks: vec![],
-            auto_resolve: None,
-            mode: ExecutionMode::Sequential,
-            span: dummy_span(),
-        };
-
-        assert_eq!(wf.find_stage("a").unwrap().agent, "agent_a");
-        assert_eq!(wf.find_stage("b").unwrap().agent, "agent_b");
-        assert!(wf.find_stage("nonexistent").is_none());
-    }
-}
