+++ src/parser/mod.rs
@@ -1,10 +1,11 @@
-use crate::ast::{
-    AgentDef, Budget, Capability, Constraint, DefaultsDef, ExecutionMode, GuardrailRule,
-    GuardrailSection, GuardrailsDef, ProviderDef, ReinFile, RouteRule, Span, Stage, ToolDef,
-    ImportDef, TypeDef, TypeExpr, TypeField, ValueExpr, WorkflowDef,
-
-};
+use crate::ast::{ReinFile, Span, ValueExpr};
 use crate::lexer::{Token, TokenKind, tokenize};
+
+mod agent_parser;
+mod import_parser;
+mod step_parser;
+mod type_parser;
+mod workflow_parser;
 
 /// Parse error with source location and message.
 #[derive(Debug, Clone, PartialEq)]
@@ -120,8 +121,6 @@ impl Parser {
 
     /// Parse a value expression: a string literal, identifier, or function
     /// call like `env("VAR_NAME")`.
-    ///
-    /// Returns a `ValueExpr` representing the parsed value.
     fn parse_value_expr(&mut self) -> Result<ValueExpr, ParseError> {
         self.skip_comments();
         let tok = self.current().clone();
@@ -211,1120 +210,6 @@ impl Parser {
         }
     }
 
-    // ── Grammar rules ────────────────────────────────────────────────────────
-
-    pub fn parse_file(&mut self) -> Result<ReinFile, ParseError> {
-        self.skip_comments();
-        let mut imports = Vec::new();
-        let mut defaults: Option<DefaultsDef> = None;
-        let mut providers = Vec::new();
-        let mut tools = Vec::new();
-        let mut agents = Vec::new();
-        let mut workflows = Vec::new();
-        let mut types = Vec::new();
-        while self.peek() != &TokenKind::Eof {
-            match self.peek() {
-                TokenKind::Import => imports.push(self.parse_import()?),
-                TokenKind::Defaults => {
-                    if defaults.is_some() {
-                        return Err(ParseError::new(
-                            "duplicate 'defaults' block (only one allowed per file)",
-                            self.current_span(),
-                        ));
-                    }
-                    defaults = Some(self.parse_defaults()?);
-                }
-                TokenKind::Provider => providers.push(self.parse_provider()?),
-                TokenKind::Tool => tools.push(self.parse_tool()?),
-                TokenKind::Agent => agents.push(self.parse_agent()?),
-                TokenKind::Workflow => workflows.push(self.parse_workflow()?),
-                TokenKind::Type => types.push(self.parse_type_def()?),
-                other => {
-                    return Err(ParseError::new(
-                        format!(
-                            "expected top-level declaration (defaults, provider, tool, agent, workflow, type), got {other}"
-                        ),
-                        self.current_span(),
-                    ));
-                }
-            }
-            self.skip_comments();
-        }
-        Ok(ReinFile {
-            imports,
-            defaults,
-            providers,
-            tools,
-            agents,
-            workflows,
-            types,
-        })
-    }
-
-    fn parse_defaults(&mut self) -> Result<DefaultsDef, ParseError> {
-        let start = self.current_span().start;
-        self.expect(&TokenKind::Defaults)?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut model: Option<ValueExpr> = None;
-        let mut budget: Option<Budget> = None;
-        let (mut seen_model, mut seen_budget) = (false, false);
-
-        loop {
-            self.skip_comments();
-            match self.peek().clone() {
-                TokenKind::RBrace => {
-                    let end = self.current_span().end;
-                    self.advance();
-                    return Ok(DefaultsDef {
-                        model,
-                        budget,
-                        span: Span::new(start, end),
-                    });
-                }
-                TokenKind::Model => {
-                    if seen_model {
-                        return Err(ParseError::new(
-                            "duplicate field 'model' in defaults block",
-                            self.current_span(),
-                        ));
-                    }
-                    seen_model = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    model = Some(self.parse_value_expr()?);
-                }
-                TokenKind::Budget => {
-                    if seen_budget {
-                        return Err(ParseError::new(
-                            "duplicate field 'budget' in defaults block",
-                            self.current_span(),
-                        ));
-                    }
-                    seen_budget = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    budget = Some(self.parse_budget()?);
-                }
-                other => {
-                    return Err(ParseError::new(
-                        format!("unexpected field in defaults block: {other}"),
-                        self.current_span(),
-                    ));
-                }
-            }
-        }
-    }
-
-    fn parse_provider(&mut self) -> Result<ProviderDef, ParseError> {
-        self.skip_comments();
-        let start = self.current_span().start;
-
-        self.expect(&TokenKind::Provider)?;
-        let (name, _) = self.expect_ident()?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut model: Option<ValueExpr> = None;
-        let mut key: Option<ValueExpr> = None;
-        let mut seen_model = false;
-        let mut seen_key = false;
-
-        loop {
-            self.skip_comments();
-            match self.peek().clone() {
-                TokenKind::RBrace => {
-                    let end = self.current_span().end;
-                    self.advance();
-                    return Ok(ProviderDef {
-                        name,
-                        model,
-                        key,
-                        span: Span::new(start, end),
-                    });
-                }
-                TokenKind::Model => {
-                    if seen_model {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'model' in provider '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_model = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    model = Some(self.parse_value_expr()?);
-                }
-                TokenKind::Key => {
-                    if seen_key {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'key' in provider '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_key = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    key = Some(self.parse_value_expr()?);
-                }
-                other => {
-                    return Err(ParseError::new(
-                        format!("unexpected field in provider '{name}': {other}"),
-                        self.current_span(),
-                    ));
-                }
-            }
-        }
-    }
-
-    fn parse_tool(&mut self) -> Result<ToolDef, ParseError> {
-        self.skip_comments();
-        let start = self.current_span().start;
-
-        self.expect(&TokenKind::Tool)?;
-        let (name, _) = self.expect_ident()?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut endpoint: Option<ValueExpr> = None;
-        let mut provider: Option<ValueExpr> = None;
-        let mut key: Option<ValueExpr> = None;
-        let mut seen_endpoint = false;
-        let mut seen_provider = false;
-        let mut seen_key = false;
-
-        loop {
-            self.skip_comments();
-            match self.peek().clone() {
-                TokenKind::RBrace => {
-                    let end = self.current_span().end;
-                    self.advance();
-                    return Ok(ToolDef {
-                        name,
-                        endpoint,
-                        provider,
-                        key,
-                        span: Span::new(start, end),
-                    });
-                }
-                TokenKind::Endpoint => {
-                    if seen_endpoint {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'endpoint' in tool '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_endpoint = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    endpoint = Some(self.parse_value_expr()?);
-                }
-                TokenKind::Provider => {
-                    if seen_provider {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'provider' in tool '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_provider = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    provider = Some(self.parse_value_expr()?);
-                }
-                TokenKind::Key => {
-                    if seen_key {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'key' in tool '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_key = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    key = Some(self.parse_value_expr()?);
-                }
-                other => {
-                    return Err(ParseError::new(
-                        format!("unexpected field in tool '{name}': {other}"),
-                        self.current_span(),
-                    ));
-                }
-            }
-        }
-    }
-
-    fn parse_agent(&mut self) -> Result<AgentDef, ParseError> {
-        self.skip_comments();
-        let start = self.current_span().start;
-
-        self.expect(&TokenKind::Agent)?;
-        let (name, _) = self.expect_ident()?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut model: Option<ValueExpr> = None;
-        let mut can: Vec<Capability> = Vec::new();
-        let mut cannot: Vec<Capability> = Vec::new();
-        let mut budget: Option<Budget> = None;
-        let mut guardrails: Option<GuardrailsDef> = None;
-
-        let (mut seen_model, mut seen_can, mut seen_cannot) = (false, false, false);
-        let (mut seen_budget, mut seen_guardrails) = (false, false);
-
-        loop {
-            self.skip_comments();
-            match self.peek().clone() {
-                TokenKind::RBrace => {
-                    let end = self.current_span().end;
-                    self.advance();
-                    return Ok(AgentDef {
-                        name,
-                        model,
-                        can,
-                        cannot,
-                        budget,
-                        guardrails,
-                        span: Span::new(start, end),
-                    });
-                }
-                TokenKind::Model => {
-                    if seen_model {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'model' in agent '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_model = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    let value = self.parse_value_expr()?;
-                    model = Some(value);
-                }
-                TokenKind::Can => {
-                    if seen_can {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'can' in agent '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_can = true;
-                    self.advance();
-                    can = self.parse_capability_list()?;
-                }
-                TokenKind::Cannot => {
-                    if seen_cannot {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'cannot' in agent '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_cannot = true;
-                    self.advance();
-                    cannot = self.parse_capability_list()?;
-                }
-                TokenKind::Budget => {
-                    if seen_budget {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'budget' in agent '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_budget = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    budget = Some(self.parse_budget()?);
-                }
-                TokenKind::Guardrails => {
-                    if seen_guardrails {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'guardrails' in agent '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_guardrails = true;
-                    guardrails = Some(self.parse_guardrails()?);
-                }
-                TokenKind::Eof => {
-                    return Err(ParseError::new(
-                        "unexpected end of file: expected `}`",
-                        self.current_span(),
-                    ));
-                }
-                other => {
-                    return Err(ParseError::new(
-                        format!("unexpected token in agent body: {other}"),
-                        self.current_span(),
-                    ));
-                }
-            }
-        }
-    }
-
-    fn parse_guardrails(&mut self) -> Result<GuardrailsDef, ParseError> {
-        let start = self.current_span().start;
-        self.expect(&TokenKind::Guardrails)?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut sections = Vec::new();
-        let mut seen_names: Vec<String> = Vec::new();
-
-        loop {
-            self.skip_comments();
-            match self.peek().clone() {
-                TokenKind::RBrace => {
-                    let end = self.current_span().end;
-                    self.advance();
-                    return Ok(GuardrailsDef {
-                        sections,
-                        span: Span::new(start, end),
-                    });
-                }
-                TokenKind::Ident(section_name) => {
-                    if seen_names.contains(&section_name) {
-                        return Err(ParseError::new(
-                            format!("duplicate guardrail section '{section_name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_names.push(section_name.clone());
-                    sections.push(self.parse_guardrail_section()?);
-                }
-                TokenKind::Eof => {
-                    return Err(ParseError::new(
-                        "unexpected end of file in guardrails block",
-                        self.current_span(),
-                    ));
-                }
-                other => {
-                    return Err(ParseError::new(
-                        format!("expected section name in guardrails, got {other}"),
-                        self.current_span(),
-                    ));
-                }
-            }
-        }
-    }
-
-    fn parse_guardrail_section(&mut self) -> Result<GuardrailSection, ParseError> {
-        let start = self.current_span().start;
-        let (name, _) = self.expect_ident()?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut rules = Vec::new();
-
-        loop {
-            self.skip_comments();
-            match self.peek().clone() {
-                TokenKind::RBrace => {
-                    let end = self.current_span().end;
-                    self.advance();
-                    return Ok(GuardrailSection {
-                        name,
-                        rules,
-                        span: Span::new(start, end),
-                    });
-                }
-                TokenKind::Ident(key) => {
-                    let rule_start = self.current_span().start;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    let (value, _) = self.expect_ident()?;
-                    let rule_end = self.current_span().start;
-                    rules.push(GuardrailRule {
-                        key,
-                        value,
-                        span: Span::new(rule_start, rule_end),
-                    });
-                }
-                TokenKind::Eof => {
-                    return Err(ParseError::new(
-                        "unexpected end of file in guardrail section",
-                        self.current_span(),
-                    ));
-                }
-                other => {
-                    return Err(ParseError::new(
-                        format!("expected rule or '}}' in guardrail section, got {other}"),
-                        self.current_span(),
-                    ));
-                }
-            }
-        }
-    }
-
-    fn parse_workflow(&mut self) -> Result<WorkflowDef, ParseError> {
-        self.skip_comments();
-        let start = self.current_span().start;
-
-        self.expect(&TokenKind::Workflow)?;
-        let (name, _) = self.expect_ident()?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut trigger: Option<String> = None;
-        let mut stages: Vec<Stage> = Vec::new();
-        let mut steps: Vec<crate::ast::StepDef> = Vec::new();
-        let mut route_blocks: Vec<crate::ast::RouteBlock> = Vec::new();
-        let mut parallel_blocks: Vec<crate::ast::ParallelBlock> = Vec::new();
-        let mut auto_resolve: Option<crate::ast::AutoResolveBlock> = None;
-        let mut seen_trigger = false;
-        let mut seen_stages = false;
-
-        loop {
-            self.skip_comments();
-            match self.peek().clone() {
-                TokenKind::RBrace => {
-                    let end = self.current_span().end;
-                    self.advance();
-
-                    let trigger = trigger.ok_or_else(|| {
-                        ParseError::new(
-                            format!("workflow '{name}' is missing required field 'trigger'"),
-                            Span::new(start, end),
-                        )
-                    })?;
-
-                    if stages.is_empty() && steps.is_empty() && route_blocks.is_empty() && parallel_blocks.is_empty() {
-                        return Err(ParseError::new(
-                            format!("workflow '{name}' must have at least one stage, step, or route block"),
-                            Span::new(start, end),
-                        ));
-                    }
-
-                    return Ok(WorkflowDef {
-                        name,
-                        trigger,
-                        stages,
-                        steps,
-                        route_blocks,
-                        parallel_blocks,
-                        auto_resolve,
-                        mode: ExecutionMode::Sequential,
-                        span: Span::new(start, end),
-                    });
-                }
-                TokenKind::Trigger => {
-                    if seen_trigger {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'trigger' in workflow '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_trigger = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    trigger = Some(self.parse_trigger_expr()?);
-                }
-                TokenKind::Stages => {
-                    if seen_stages {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'stages' in workflow '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_stages = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    stages = self.parse_stage_list()?;
-                }
-                TokenKind::Step => {
-                    steps.push(self.parse_step()?);
-                }
-                TokenKind::Route => {
-                    route_blocks.push(self.parse_route_block()?);
-                }
-                TokenKind::Parallel => {
-                    parallel_blocks.push(self.parse_parallel_block()?);
-                }
-                TokenKind::Auto => {
-                    if auto_resolve.is_some() {
-                        return Err(ParseError::new(
-                            format!("duplicate 'auto resolve when' in workflow '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    auto_resolve = Some(self.parse_auto_resolve()?);
-                }
-                TokenKind::Eof => {
-                    return Err(ParseError::new(
-                        "unexpected end of file: expected `}`",
-                        self.current_span(),
-                    ));
-                }
-                other => {
-                    return Err(ParseError::new(
-                        format!("unexpected token in workflow body: {other}"),
-                        self.current_span(),
-                    ));
-                }
-            }
-        }
-    }
-
-    fn parse_stage_list(&mut self) -> Result<Vec<Stage>, ParseError> {
-        self.expect(&TokenKind::LBracket)?;
-        let mut stages = Vec::new();
-        loop {
-            self.skip_comments();
-            match self.peek() {
-                TokenKind::RBracket => {
-                    self.advance();
-                    break;
-                }
-                TokenKind::Eof => {
-                    return Err(ParseError::new(
-                        "unexpected end of file: expected `]`",
-                        self.current_span(),
-                    ));
-                }
-                _ => {
-                    let stage_start = self.current_span().start;
-                    let (agent_name, _) = self.expect_ident()?;
-                    let end = self.last_consumed_end;
-
-                    stages.push(Stage {
-                        name: agent_name.clone(),
-                        agent: agent_name,
-                        route: RouteRule::Next,
-                        span: Span::new(stage_start, end),
-                    });
-
-                    // Optional comma separator
-                    if self.peek() == &TokenKind::Comma {
-                        self.advance();
-                    }
-                }
-            }
-        }
-        Ok(stages)
-    }
-
-    /// Parse a `step <name> { agent: <ident> goal: <text> }` block.
-    #[allow(clippy::too_many_lines)]
-    fn parse_step(&mut self) -> Result<crate::ast::StepDef, ParseError> {
-        self.skip_comments();
-        let start = self.current_span().start;
-
-        self.expect(&TokenKind::Step)?;
-        let (name, _) = self.expect_ident()?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut agent: Option<String> = None;
-        let mut goal: Option<String> = None;
-        let mut output_constraints: Vec<(String, TypeExpr)> = Vec::new();
-        let mut when: Option<crate::ast::WhenExpr> = None;
-        let mut on_failure: Option<crate::ast::RetryPolicy> = None;
-        let mut seen_agent = false;
-        let mut seen_goal = false;
-
-        loop {
-            self.skip_comments();
-            match self.peek().clone() {
-                TokenKind::RBrace => {
-                    let end = self.current_span().end;
-                    self.advance();
-
-                    let agent = agent.ok_or_else(|| {
-                        ParseError::new(
-                            format!("step '{name}' is missing required field 'agent'"),
-                            Span::new(start, end),
-                        )
-                    })?;
-
-                    return Ok(crate::ast::StepDef {
-                        name,
-                        agent,
-                        goal,
-                        output_constraints,
-                        when,
-                        on_failure,
-                        span: Span::new(start, end),
-                    });
-                }
-                TokenKind::Agent => {
-                    if seen_agent {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'agent' in step '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_agent = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    let (value, _) = self.expect_ident()?;
-                    agent = Some(value);
-                }
-                TokenKind::Goal => {
-                    if seen_goal {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'goal' in step '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    seen_goal = true;
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    // Goal is a string literal or plain text value
-                    let value = self.parse_value_expr()?;
-                    goal = Some(match value {
-                        ValueExpr::Literal(s) => s,
-                        ValueExpr::EnvRef { span, .. } => {
-                            return Err(ParseError::new(
-                                "goal must be a string literal, not env()",
-                                span,
-                            ));
-                        }
-                    });
-                }
-                TokenKind::On => {
-                    if on_failure.is_some() {
-                        return Err(ParseError::new(
-                            format!("duplicate 'on failure' in step '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    on_failure = Some(self.parse_retry_policy()?);
-                }
-                TokenKind::When => {
-                    if when.is_some() {
-                        return Err(ParseError::new(
-                            format!("duplicate field 'when' in step '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                    self.advance();
-                    self.expect(&TokenKind::Colon)?;
-                    when = Some(self.parse_when_expr()?);
-                }
-                TokenKind::Ident(ref field_name)
-                    if self.peek_at(1).is_some_and(|t| *t == TokenKind::Colon) =>
-                {
-                    let field_name = field_name.clone();
-                    self.advance(); // consume ident
-                    self.expect(&TokenKind::Colon)?;
-                    if *self.peek() == TokenKind::One {
-                        let type_expr = self.parse_one_of()?;
-                        output_constraints.push((field_name, type_expr));
-                    } else {
-                        return Err(ParseError::new(
-                            format!("unexpected field '{field_name}' in step '{name}'"),
-                            self.current_span(),
-                        ));
-                    }
-                }
-                TokenKind::Eof => {
-                    return Err(ParseError::new(
-                        "unexpected end of file: expected `}`",
-                        self.current_span(),
-                    ));
-                }
-                other => {
-                    return Err(ParseError::new(
-                        format!("unexpected token in step '{name}': {other}"),
-                        self.current_span(),
-                    ));
-                }
-            }
-        }
-    }
-
-    /// Parse `on failure: retry N strategy then action`.
-    fn parse_retry_policy(&mut self) -> Result<crate::ast::RetryPolicy, ParseError> {
-        self.expect(&TokenKind::On)?;
-        self.expect(&TokenKind::Failure)?;
-        self.expect(&TokenKind::Colon)?;
-        self.expect(&TokenKind::Retry)?;
-
-        // Parse retry count
-        let max_retries = match self.peek().clone() {
-            TokenKind::Number(n) => {
-                let val = n.parse::<u32>().map_err(|_| {
-                    ParseError::new(
-                        format!("invalid retry count: {n}"),
-                        self.current_span(),
-                    )
-                })?;
-                self.advance();
-                val
-            }
-            other => {
-                return Err(ParseError::new(
-                    format!("expected retry count (number), got {other}"),
-                    self.current_span(),
-                ));
-            }
-        };
-
-        // Parse backoff strategy
-        let backoff = match self.peek() {
-            TokenKind::Exponential => {
-                self.advance();
-                crate::ast::BackoffStrategy::Exponential
-            }
-            TokenKind::Linear => {
-                self.advance();
-                crate::ast::BackoffStrategy::Linear
-            }
-            TokenKind::Fixed => {
-                self.advance();
-                crate::ast::BackoffStrategy::Fixed
-            }
-            other => {
-                return Err(ParseError::new(
-                    format!("expected backoff strategy (exponential, linear, fixed), got {other}"),
-                    self.current_span(),
-                ));
-            }
-        };
-
-        self.expect(&TokenKind::Then)?;
-
-        // Parse failure action
-        let then = match self.peek() {
-            TokenKind::Escalate => {
-                self.advance();
-                crate::ast::FailureAction::Escalate
-            }
-            TokenKind::Ident(name) => {
-                let name = name.clone();
-                self.advance();
-                crate::ast::FailureAction::Step(name)
-            }
-            other => {
-                return Err(ParseError::new(
-                    format!("expected failure action (escalate or step name), got {other}"),
-                    self.current_span(),
-                ));
-            }
-        };
-
-        Ok(crate::ast::RetryPolicy {
-            max_retries,
-            backoff,
-            then,
-        })
-    }
-
-    /// Parse a `when` expression: `field op value [or|and field op value ...]`.
-    fn parse_when_expr(&mut self) -> Result<crate::ast::WhenExpr, ParseError> {
-        let first = self.parse_when_comparison()?;
-        let mut expr = crate::ast::WhenExpr::Comparison(first);
-
-        loop {
-            self.skip_comments();
-            match self.peek() {
-                TokenKind::Or => {
-                    self.advance();
-                    let next = self.parse_when_comparison()?;
-                    expr = match expr {
-                        crate::ast::WhenExpr::Or(mut parts) => {
-                            parts.push(crate::ast::WhenExpr::Comparison(next));
-                            crate::ast::WhenExpr::Or(parts)
-                        }
-                        other => crate::ast::WhenExpr::Or(vec![
-                            other,
-                            crate::ast::WhenExpr::Comparison(next),
-                        ]),
-                    };
-                }
-                TokenKind::And => {
-                    self.advance();
-                    let next = self.parse_when_comparison()?;
-                    expr = match expr {
-                        crate::ast::WhenExpr::And(mut parts) => {
-                            parts.push(crate::ast::WhenExpr::Comparison(next));
-                            crate::ast::WhenExpr::And(parts)
-                        }
-                        other => crate::ast::WhenExpr::And(vec![
-                            other,
-                            crate::ast::WhenExpr::Comparison(next),
-                        ]),
-                    };
-                }
-                _ => break,
-            }
-        }
-        Ok(expr)
-    }
-
-    /// Parse a single comparison: `field op value`.
-    fn parse_when_comparison(&mut self) -> Result<crate::ast::WhenComparison, ParseError> {
-        let (field, _) = self.expect_ident()?;
-
-        let op = match self.peek() {
-            TokenKind::Lt => crate::ast::CompareOp::Lt,
-            TokenKind::Gt => crate::ast::CompareOp::Gt,
-            TokenKind::LtEq => crate::ast::CompareOp::LtEq,
-            TokenKind::GtEq => crate::ast::CompareOp::GtEq,
-            other => {
-                return Err(ParseError::new(
-                    format!("expected comparison operator (<, >, <=, >=), got {other}"),
-                    self.current_span(),
-                ));
-            }
-        };
-        self.advance();
-
-        let value = self.parse_when_value()?;
-
-        Ok(crate::ast::WhenComparison { field, op, value })
-    }
-
-    /// Parse a when value: number, percent, currency, or ident.
-    fn parse_when_value(&mut self) -> Result<crate::ast::WhenValue, ParseError> {
-        match self.peek().clone() {
-            TokenKind::Number(n) => {
-                let n = n.clone();
-                self.advance();
-                // Check for trailing %
-                if *self.peek() == TokenKind::Percent {
-                    self.advance();
-                    Ok(crate::ast::WhenValue::Percent(n))
-                } else {
-                    Ok(crate::ast::WhenValue::Number(n))
-                }
-            }
-            TokenKind::Currency { symbol, amount } => {
-                self.advance();
-                Ok(crate::ast::WhenValue::Currency { symbol, amount })
-            }
-            TokenKind::Ident(name) => {
-                let name = name.clone();
-                self.advance();
-                Ok(crate::ast::WhenValue::Ident(name))
-            }
-            other => Err(ParseError::new(
-                format!("expected value (number, percentage, currency, or identifier), got {other}"),
-                self.current_span(),
-            )),
-        }
-    }
-
-    /// Parse `auto resolve when { condition, condition, ... }`.
-    fn parse_auto_resolve(&mut self) -> Result<crate::ast::AutoResolveBlock, ParseError> {
-        let start = self.current_span().start;
-        self.expect(&TokenKind::Auto)?;
-        self.expect(&TokenKind::Resolve)?;
-        self.expect(&TokenKind::When)?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut conditions = Vec::new();
-        loop {
-            self.skip_comments();
-            if *self.peek() == TokenKind::RBrace {
-                break;
-            }
-            conditions.push(self.parse_auto_resolve_condition()?);
-            // Optional comma separator
-            if *self.peek() == TokenKind::Comma {
-                self.advance();
-            }
-        }
-        let end = self.current_span().end;
-        self.expect(&TokenKind::RBrace)?;
-
-        Ok(crate::ast::AutoResolveBlock {
-            conditions,
-            span: Span::new(start, end),
-        })
-    }
-
-    /// Parse a single auto-resolve condition.
-    ///
-    /// Either a comparison (`confidence > 90%`) or a membership check
-    /// (`action is one of [order_status, tracking]`).
-    fn parse_auto_resolve_condition(
-        &mut self,
-    ) -> Result<crate::ast::AutoResolveCondition, ParseError> {
-        let (field, _) = self.expect_ident()?;
-
-        // Check if next is `is one of [...]`
-        if *self.peek() == TokenKind::Is {
-            self.advance(); // is
-            let type_expr = self.parse_one_of()?;
-            match type_expr {
-                TypeExpr::OneOf { variants, .. } => {
-                    Ok(crate::ast::AutoResolveCondition::IsOneOf { field, variants })
-                }
-                _ => Err(ParseError::new(
-                    "expected 'one of [...]' after 'is'",
-                    self.current_span(),
-                )),
-            }
-        } else {
-            // It's a comparison
-            let op = match self.peek() {
-                TokenKind::Lt => crate::ast::CompareOp::Lt,
-                TokenKind::Gt => crate::ast::CompareOp::Gt,
-                TokenKind::LtEq => crate::ast::CompareOp::LtEq,
-                TokenKind::GtEq => crate::ast::CompareOp::GtEq,
-                other => {
-                    return Err(ParseError::new(
-                        format!("expected comparison operator or 'is', got {other}"),
-                        self.current_span(),
-                    ));
-                }
-            };
-            self.advance();
-            let value = self.parse_when_value()?;
-            Ok(crate::ast::AutoResolveCondition::Comparison(
-                crate::ast::WhenComparison { field, op, value },
-            ))
-        }
-    }
-
-    /// Parse a `parallel { step a {...} step b {...} }` block.
-    fn parse_parallel_block(&mut self) -> Result<crate::ast::ParallelBlock, ParseError> {
-        let start = self.current_span().start;
-        self.expect(&TokenKind::Parallel)?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut steps = Vec::new();
-        loop {
-            self.skip_comments();
-            if *self.peek() == TokenKind::RBrace {
-                break;
-            }
-            steps.push(self.parse_step()?);
-        }
-        let end = self.current_span().end;
-        self.expect(&TokenKind::RBrace)?;
-
-        if steps.is_empty() {
-            return Err(ParseError::new(
-                "parallel block must contain at least one step",
-                Span::new(start, end),
-            ));
-        }
-
-        Ok(crate::ast::ParallelBlock {
-            steps,
-            span: Span::new(start, end),
-        })
-    }
-
-    /// Parse a `route on <field_path> { pattern -> step name { ... }, ... }` block.
-    fn parse_route_block(&mut self) -> Result<crate::ast::RouteBlock, ParseError> {
-        let start = self.current_span().start;
-        self.expect(&TokenKind::Route)?;
-        self.expect(&TokenKind::On)?;
-
-        // Parse dot-separated field path (keywords allowed as segments)
-        let (first, _) = self.expect_ident()?;
-        let mut path = first;
-        while *self.peek() == TokenKind::Dot {
-            self.advance(); // .
-            // Allow keywords as path segments
-            let segment = self.expect_ident_or_keyword()?;
-            path.push('.');
-            path.push_str(&segment);
-        }
-
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut arms = Vec::new();
-        loop {
-            self.skip_comments();
-            if *self.peek() == TokenKind::RBrace {
-                break;
-            }
-            let arm_start = self.current_span().start;
-            let pattern = if *self.peek() == TokenKind::Underscore {
-                self.advance();
-                crate::ast::RoutePattern::Wildcard
-            } else {
-                let (val, _) = self.expect_ident()?;
-                crate::ast::RoutePattern::Value(val)
-            };
-            self.expect(&TokenKind::Arrow)?;
-            let step = self.parse_step()?;
-            let arm_end = self.current_span().start;
-            arms.push(crate::ast::RouteArm {
-                pattern,
-                step,
-                span: Span::new(arm_start, arm_end),
-            });
-        }
-        let end = self.current_span().end;
-        self.expect(&TokenKind::RBrace)?;
-
-        Ok(crate::ast::RouteBlock {
-            field_path: path,
-            arms,
-            span: Span::new(start, end),
-        })
-    }
-
-    /// Parse an import declaration.
-    ///
-    /// Supports:
-    /// - `import { name1, name2 } from "./path.rein"`
-    /// - `import all from "./dir/"`
-    /// - `import from @scope/name`
-    fn parse_import(&mut self) -> Result<ImportDef, ParseError> {
-        let start = self.current_span().start;
-        self.expect(&TokenKind::Import)?;
-
-        match self.peek().clone() {
-            // import { name, ... } from "path"
-            TokenKind::LBrace => {
-                self.advance(); // {
-                let mut names = Vec::new();
-                loop {
-                    self.skip_comments();
-                    if *self.peek() == TokenKind::RBrace {
-                        break;
-                    }
-                    let (name, _) = self.expect_ident()?;
-                    names.push(name);
-                    self.skip_comments();
-                    if *self.peek() == TokenKind::Comma {
-                        self.advance();
-                    }
-                }
-                self.expect(&TokenKind::RBrace)?;
-                self.expect(&TokenKind::From)?;
-                let source = self.expect_string()?;
-                let end = self.current_span().start;
-                Ok(ImportDef::Named {
-                    names,
-                    source,
-                    span: Span::new(start, end),
-                })
-            }
-            // import all from "path"
-            TokenKind::All => {
-                self.advance(); // all
-                self.expect(&TokenKind::From)?;
-                let source = self.expect_string()?;
-                let end = self.current_span().start;
-                Ok(ImportDef::Glob {
-                    source,
-                    span: Span::new(start, end),
-                })
-            }
-            // import from @scope/name
-            TokenKind::From => {
-                self.advance(); // from
-                self.expect(&TokenKind::At)?;
-                let (scope, _) = self.expect_ident()?;
-                // Expect '/' separator — we'll use Ident since / isn't a token
-                // Actually, we need a slash. Let me handle this differently.
-                // For @scope/name, we'll read the rest as ident after consuming
-                // what we can. Since '/' isn't a token, we'll expect the name
-                // to be separated by checking for Dot or another convention.
-                self.expect(&TokenKind::Slash)?;
-                let (name, _) = self.expect_ident()?;
-                let end = self.current_span().start;
-                Ok(ImportDef::Registry {
-                    scope,
-                    name,
-                    span: Span::new(start, end),
-                })
-            }
-            other => Err(ParseError::new(
-                format!("expected '{{', 'all', or 'from' after 'import', got {other}"),
-                self.current_span(),
-            )),
-        }
-    }
-
     /// Expect an identifier or keyword token, returning the text.
     /// Useful in contexts where keywords are valid (e.g. field paths).
     fn expect_ident_or_keyword(&mut self) -> Result<String, ParseError> {
@@ -1369,175 +254,6 @@ impl Parser {
         }
     }
 
-    /// Parse a `type Name { field: type_expr, ... }` definition.
-    fn parse_type_def(&mut self) -> Result<TypeDef, ParseError> {
-        let start = self.current_span().start;
-        self.expect(&TokenKind::Type)?;
-        let (name, _) = self.expect_ident()?;
-        self.expect(&TokenKind::LBrace)?;
-
-        let mut fields = Vec::new();
-        loop {
-            self.skip_comments();
-            if *self.peek() == TokenKind::RBrace {
-                break;
-            }
-            let field_start = self.current_span().start;
-            let (field_name, _) = self.expect_ident()?;
-            self.expect(&TokenKind::Colon)?;
-            let type_expr = self.parse_type_expr()?;
-            let field_end = self.current_span().start;
-            fields.push(TypeField {
-                name: field_name,
-                type_expr,
-                span: Span::new(field_start, field_end),
-            });
-        }
-        let end = self.current_span().end;
-        self.expect(&TokenKind::RBrace)?;
-
-        Ok(TypeDef {
-            name,
-            fields,
-            span: Span::new(start, end),
-        })
-    }
-
-    /// Parse a type expression: named type, array, one of, or range.
-    fn parse_type_expr(&mut self) -> Result<TypeExpr, ParseError> {
-        self.skip_comments();
-        match self.peek().clone() {
-            TokenKind::One => self.parse_one_of(),
-            TokenKind::Number(n) => {
-                let min = n.clone();
-                self.advance();
-                self.expect(&TokenKind::DotDot)?;
-                match self.peek().clone() {
-                    TokenKind::Number(max) => {
-                        let max = max.clone();
-                        self.advance();
-                        Ok(TypeExpr::Range { min, max })
-                    }
-                    other => Err(ParseError::new(
-                        format!("expected number after '..', got {other}"),
-                        self.current_span(),
-                    )),
-                }
-            }
-            TokenKind::Ident(name) => {
-                let name = name.clone();
-                self.advance();
-                // Check for array syntax: Type[]
-                let array = if *self.peek() == TokenKind::LBracket
-                    && self.peek_at(1).is_some_and(|t| *t == TokenKind::RBracket)
-                {
-                    self.advance(); // [
-                    self.advance(); // ]
-                    true
-                } else {
-                    false
-                };
-                Ok(TypeExpr::Named { name, array })
-            }
-            other => Err(ParseError::new(
-                format!("expected type expression, got {other}"),
-                self.current_span(),
-            )),
-        }
-    }
-
-    /// Parse `one of [a, b, c]` type expression.
-    fn parse_one_of(&mut self) -> Result<TypeExpr, ParseError> {
-        let start = self.current_span().start;
-        self.expect(&TokenKind::One)?;
-        self.expect(&TokenKind::Of)?;
-        self.expect(&TokenKind::LBracket)?;
-
-        let mut variants = Vec::new();
-        loop {
-            self.skip_comments();
-            if *self.peek() == TokenKind::RBracket {
-                break;
-            }
-            let (variant, _) = self.expect_ident()?;
-            variants.push(variant);
-            self.skip_comments();
-            if *self.peek() == TokenKind::Comma {
-                self.advance();
-            }
-        }
-        let end = self.current_span().end;
-        self.expect(&TokenKind::RBracket)?;
-
-        if variants.is_empty() {
-            return Err(ParseError::new(
-                "one of requires at least one variant",
-                Span::new(start, end),
-            ));
-        }
-
-        Ok(TypeExpr::OneOf {
-            variants,
-            span: Span::new(start, end),
-        })
-    }
-
-    fn parse_capability_list(&mut self) -> Result<Vec<Capability>, ParseError> {
-        self.expect(&TokenKind::LBracket)?;
-        let mut caps = Vec::new();
-        loop {
-            self.skip_comments();
-            match self.peek() {
-                TokenKind::RBracket => {
-                    self.advance();
-                    break;
-                }
-                TokenKind::Eof => {
-                    return Err(ParseError::new(
-                        "unexpected end of file: expected `]`",
-                        self.current_span(),
-                    ));
-                }
-                _ => caps.push(self.parse_capability()?),
-            }
-        }
-        Ok(caps)
-    }
-
-    fn parse_capability(&mut self) -> Result<Capability, ParseError> {
-        let start = self.current_span().start;
-        let (namespace, _) = self.expect_ident()?;
-        self.expect(&TokenKind::Dot)?;
-        let (action, _) = self.expect_ident()?;
-
-        // optional `up to $<amount>`
-        let constraint = if self.peek() == &TokenKind::Up {
-            self.advance();
-            self.expect(&TokenKind::To)?;
-            let (amount, symbol, _) = self.expect_currency()?;
-            let currency = match symbol {
-                '€' => "EUR",
-                '£' => "GBP",
-                '¥' => "JPY",
-                _ => "USD",
-            };
-            Some(Constraint::MonetaryCap {
-                amount,
-                currency: currency.to_string(),
-            })
-        } else {
-            None
-        };
-
-        let end = self.last_consumed_end;
-        Ok(Capability {
-            namespace,
-            action,
-            constraint,
-            span: Span::new(start, end),
-        })
-    }
-
     fn expect_currency(&mut self) -> Result<(u64, char, Span), ParseError> {
         self.skip_comments();
         let tok = self.current().clone();
@@ -1553,53 +269,54 @@ impl Parser {
         }
     }
 
-    fn parse_budget(&mut self) -> Result<Budget, ParseError> {
-        let start = self.current_span().start;
-        let (amount, symbol, _) = self.expect_currency()?;
-        self.expect(&TokenKind::Per)?;
-        let (unit, _) = self.expect_ident()?;
-        let end = self.last_consumed_end;
-        let currency = match symbol {
-            '€' => "EUR",
-            '£' => "GBP",
-            '¥' => "JPY",
-            _ => "USD",
-        };
-        Ok(Budget {
-            amount,
-            currency: currency.to_string(),
-            unit,
-            span: Span::new(start, end),
-        })
-    }
-    /// Parse a trigger expression: either a string literal or a sequence of
-    /// identifiers (e.g., `new ticket in zendesk`).
-    fn parse_trigger_expr(&mut self) -> Result<String, ParseError> {
+    // ── Grammar rules ────────────────────────────────────────────────────────
+
+    fn parse_file(&mut self) -> Result<ReinFile, ParseError> {
         self.skip_comments();
-        match self.peek().clone() {
-            TokenKind::StringLiteral(s) => {
-                self.advance();
-                Ok(s)
-            }
-            TokenKind::Ident(_) => {
-                let mut parts = Vec::new();
-                while let TokenKind::Ident(word) = self.peek().clone() {
-                    parts.push(word);
-                    self.advance();
+        let mut imports = Vec::new();
+        let mut defaults = None;
+        let mut providers = Vec::new();
+        let mut tools = Vec::new();
+        let mut agents = Vec::new();
+        let mut workflows = Vec::new();
+        let mut types = Vec::new();
+        while self.peek() != &TokenKind::Eof {
+            match self.peek() {
+                TokenKind::Import => imports.push(self.parse_import()?),
+                TokenKind::Defaults => {
+                    if defaults.is_some() {
+                        return Err(ParseError::new(
+                            "duplicate 'defaults' block (only one allowed per file)",
+                            self.current_span(),
+                        ));
+                    }
+                    defaults = Some(self.parse_defaults()?);
                 }
-                if parts.is_empty() {
+                TokenKind::Provider => providers.push(self.parse_provider()?),
+                TokenKind::Tool => tools.push(self.parse_tool()?),
+                TokenKind::Agent => agents.push(self.parse_agent()?),
+                TokenKind::Workflow => workflows.push(self.parse_workflow()?),
+                TokenKind::Type => types.push(self.parse_type_def()?),
+                other => {
                     return Err(ParseError::new(
-                        "expected trigger expression",
+                        format!(
+                            "expected top-level declaration (defaults, provider, tool, agent, workflow, type), got {other}"
+                        ),
                         self.current_span(),
                     ));
                 }
-                Ok(parts.join(" "))
             }
-            _ => Err(ParseError::new(
-                format!("expected trigger expression, got {}", self.peek()),
-                self.current_span(),
-            )),
+            self.skip_comments();
         }
+        Ok(ReinFile {
+            imports,
+            defaults,
+            providers,
+            tools,
+            agents,
+            workflows,
+            types,
+        })
     }
 }
